refactor: split netcheck webview into modules and dedupe diagnostic command flow
- Replace the monolithic render.ts/index.ts in packages/netcheck with
  focused modules: format, health, connectivity, regions, dom, page
- Extract the shared CLI-diagnostic command flow (progress, telemetry,
  error handling) from runSpeedTest/runNetcheck into
  src/command/diagnosticFlow.ts
- Add an emptyIfNull schema helper and remove an Object.fromEntries
  cast in the netcheck report schema
- Simplify NetcheckPanelFactory.show to take NetcheckData directly
- Split render.test.ts into per-module test files with a shared fixture

Rendering is unchanged: harness screenshots are byte-identical before
and after across themes, states, and widths.
