+++ src/commands.ts
@@ -3,13 +3,13 @@ import * as os from "node:os";
 import * as path from "node:path";
 import * as semver from "semver";
 import * as vscode from "vscode";
-import { ZodError } from "zod";
 
 import {
 	createWorkspaceIdentifier,
 	extractAgents,
 	workspaceStatusLabel,
 } from "./api/api-helper";
+import { runDiagnosticCli } from "./command/diagnosticFlow";
 import * as cliExec from "./core/cliExec";
 import { CertificateError } from "./error/certificateError";
 import { toError } from "./error/errorUtils";
@@ -314,8 +314,12 @@ export class Commands {
 		const seconds = Number(input.trim());
 		telemetry.setRequestedDuration(seconds);
 
-		const result = await withCancellableProgress(
-			async ({ signal, progress }) => {
+		await runDiagnosticCli({
+			telemetry,
+			logger: this.logger,
+			name: "Speed test",
+			progressTitle: `Running speed test for ${workspaceId}`,
+			exec: async ({ signal, progress }) => {
 				progress.report({ message: "Connecting..." });
 				const env = await this.resolveCliEnv(client);
 
@@ -338,49 +342,16 @@ export class Commands {
 					stopProgress();
 				}
 			},
-			{
-				location: vscode.ProgressLocation.Notification,
-				title: `Running speed test for ${workspaceId}`,
-				cancellable: true,
+			display: (rawJson) => {
+				const parsed = parseSpeedtestResult(rawJson);
+				telemetry.succeedSpeedtest(parsed);
+				this.speedtestPanelFactory.show({
+					result: parsed,
+					rawJson,
+					workspaceId,
+				});
 			},
-		);
-
-		if (!result.ok) {
-			if (result.cancelled) {
-				telemetry.abort("progress");
-				return;
-			}
-			telemetry.error();
-			this.logger.error("Speed test failed", result.error);
-			vscode.window.showErrorMessage(
-				`Speed test failed: ${toError(result.error).message}`,
-			);
-			return;
-		}
-
-		try {
-			const parsed = parseSpeedtestResult(result.value);
-			telemetry.succeedSpeedtest(parsed);
-			this.speedtestPanelFactory.show({
-				result: parsed,
-				rawJson: result.value,
-				workspaceId,
-			});
-		} catch (err) {
-			if (err instanceof ZodError || err instanceof SyntaxError) {
-				telemetry.error("parse_error");
-				this.logger.error("Failed to parse speedtest output", err);
-				vscode.window.showErrorMessage(
-					"Speed test output did not match the expected format. Check `Output > Coder` for details.",
-				);
-				return;
-			}
-			telemetry.error();
-			this.logger.error("Failed to display speedtest results", err);
-			vscode.window.showErrorMessage(
-				`Speed test returned unexpected output: ${toError(err).message}`,
-			);
-		}
+		});
 	}
 
 	/**
@@ -407,8 +378,12 @@ export class Commands {
 		}
 		const host = toSafeHost(baseUrl);
 
-		const result = await withCancellableProgress(
-			async ({ signal, progress }) => {
+		await runDiagnosticCli({
+			telemetry,
+			logger: this.logger,
+			name: "Network check",
+			progressTitle: `Running network check for ${host}`,
+			exec: async ({ signal, progress }) => {
 				progress.report({ message: "Resolving CLI..." });
 				const env = await this.resolveCliEnv(client);
 				progress.report({
@@ -416,49 +391,12 @@ export class Commands {
 				});
 				return await cliExec.netcheck(env, signal);
 			},
-			{
-				location: vscode.ProgressLocation.Notification,
-				title: `Running network check for ${host}`,
-				cancellable: true,
+			display: (rawJson) => {
+				const report = parseNetcheckReport(rawJson);
+				telemetry.succeedNetcheck(report);
+				this.netcheckPanelFactory.show({ host, report }, rawJson);
 			},
-		);
-
-		if (!result.ok) {
-			if (result.cancelled) {
-				telemetry.abort("progress");
-				return;
-			}
-			telemetry.error();
-			this.logger.error("Network check failed", result.error);
-			vscode.window.showErrorMessage(
-				`Network check failed: ${toError(result.error).message}`,
-			);
-			return;
-		}
-
-		try {
-			const report = parseNetcheckReport(result.value);
-			telemetry.succeedNetcheck(report);
-			this.netcheckPanelFactory.show({
-				report,
-				rawJson: result.value,
-				host,
-			});
-		} catch (err) {
-			if (err instanceof ZodError || err instanceof SyntaxError) {
-				telemetry.error("parse_error");
-				this.logger.error("Failed to parse netcheck output", err);
-				vscode.window.showErrorMessage(
-					"Network check output did not match the expected format. Check `Output > Coder` for details.",
-				);
-				return;
-			}
-			telemetry.error();
-			this.logger.error("Failed to display netcheck results", err);
-			vscode.window.showErrorMessage(
-				`Network check returned unexpected output: ${toError(err).message}`,
-			);
-		}
+		});
 	}
 
 	public async supportBundle(item?: OpenableTreeItem): Promise<void> {
